Move SurrealDB server health check into its own file
